BoardSquare how has coveredby fields to help with moveset gen etc.
Instead of the insane bitmasking of a nother dict, I added some fields to all boardsquares.  Now when generating movesets we set status of squares if they are covered by a piece (for calcing checkmate, et al.).  Doesn't work for pawns yet, since they have an unusal mode of capturing.